complete refactorization of home_screen.dart
