+++ lib/core/presentation/routes.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 //my imports
 import 'package:kabod_app/screens/auth/screens/login_screens_controller.dart';
 import 'package:kabod_app/screens/auth/screens/reset_password_screen.dart';
-import 'package:kabod_app/screens/home/screens/home_screen.dart';
+import 'package:kabod_app/screens/home/home_screen.dart';
 
 class AppRoutes {
   static const String loginRoute = '/';
