+++ lib/screens/auth/screens/login_screens_controller.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'package:kabod_app/screens/auth/model/user_model.dart';
 import 'package:kabod_app/screens/auth/screens/login_screen.dart';
 import 'package:kabod_app/screens/auth/screens/splash.dart';
-import 'package:kabod_app/screens/home/screens/home_screen.dart';
+import 'package:kabod_app/screens/home//home_screen.dart';
 
 class HomePage extends StatelessWidget {
   @override
