+++ lib/screens/commons/dividers.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class DividerBig extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return DividerBig();
+  }
+}
+
+class DividerMedium extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return DividerMedium();
+  }
+}
+
+class DividerSmall extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return DividerSmall();
+  }
+}
